requirements.txt
numpy~=1.20.3
scipy~=1.7.0
pandas~=1.2.5
PyYAML~=5.3.1
numba~=0.54.0
APScheduler~=3.6.3
SQLAlchemy~=1.3.17
AMQPStorm~=2.7.2
scikit-learn~=1.0.2
tqdm~=4.64.0
python-dateutil~=2.8.2
matplotlib
seaborn
shodan
-e git+https://github.com/CESNET/dp3.git@da829dd8#egg=dp3
jupyter